sprint-3/step-1 feat: вынести утилиты в utils и убрать console.log
