+++ src/pages/AuthPage/AuthPage.jsx
@@ -1,8 +1,6 @@
 import React from 'react'
-import { useSelector } from 'react-redux'
 import { Redirect } from 'react-router-dom'
 import styles from './AuthPage.css'
-import AppHeader from '../../components/AppHeader/AppHeader'
 import AuthForm from '../../components/AuthForm/AuthForm'
 import { getCookie } from "../../utils/coockies"
